Fix response status handling and remove unused imports
Updated `CampTypeController` to properly handle response status without unnecessary casting. Removed unused imports in multiple repository files to clean up the codebase and enhance maintainability. No functional behavior is impacted.

+++ src/main/java/com/group2/glamping/controller/CampTypeController.java
@@ -72,7 +72,7 @@ public class CampTypeController {
     @DeleteMapping("/{campTypeId}")
     public ResponseEntity<BaseResponse> deleteCampType(@PathVariable int campTypeId) {
         BaseResponse response = campTypeService.softDeleteCampType(campTypeId);
-        return ResponseEntity.status((int) response.getStatusCode()).body(response);
+        return ResponseEntity.status(response.getStatusCode()).body(response);
     }
 }
